+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Member and channel list are now resizable
+
 ## [0.4.2] - 2026-02-13
 
 - Screenshare and video support